chore: move setCookie function to utils folder
