+++ src/app/_lib/utils/setCookie.ts
@@ -1,6 +1,8 @@
 "use server";
 import { cookies } from "next/headers";
 
-export async function setCookie(key: string, value: string) {
+const setCookie = (key: string, value: string) => {
   cookies().set(key, value);
-}
+};
+
+export default setCookie;
